+++ src/templates/index.php
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="%LANGUAGE%">
+
+	<head>
+
+		<meta charset="utf-8"/>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta name="robots" content="%ROBOTS%" />
+		<meta http-equiv="content-type" content="text/html; charset=UTF-8" />
+		<meta http-equiv="content-language" content="%LANGUAGE%" />
+		<meta http-equiv="X-UA-Compatible" content="IE=9; IE=8;" />
+
+		<title>%SITE_NAME%</title>
+
+		<!--%META_DATA%-->
+		<!--%FAVICON%-->
+
+		<link media="screen" rel="stylesheet" type="text/css" href="%ROOT%/libs/bootstrap/css/bootstrap.min.css" />
+		<link media="screen" rel="stylesheet" type="text/css" href="%ROOT%/libs/font-awesome/css/font-awesome.min.css" />
+		<link media="screen" rel="stylesheet" type="text/css" href="%ROOT%/templates/assets/css/index.css" />
+
+		<!--%ADDITIONNAL_CSS%-->
+
+	</head>
+
+	<body>
+		<div class="container">
+			<div class="page-header"><h1>%SITE_NAME%</h1></div>
+
+			%CONTENT%
+
+		</div>
+	</body>
+</html>